Adds `_build_sampler` method to the `BaseModel` class in `base/base_model.py` since the tensorflow operations created in this method are universal to all models.
Additionally, updates the `_create_inputs` method to include the
placeholder `eps_ph` that allows for the step size `eps` to be changed
dynamically when running inference.

To specify a custom value to use for the step size, pass `--eps` as a
command line flag when running inference (i.e. either `gmm_inference.py`
or `gauge_model_inference.py`.

NOTE: The logic responsible for actually changing/setting the step size
(for the GMM) can be found in the `set_eps` method of `gmm_inference.py`.