Removes redundant code from the `build` method of both the `GaugeModel` and `GaussianMixtureModel` objects.
Explicitly, the code responsible for creating the operations `x_out`,
`px`, etc. needed to run the L2HMC sampler is common to all models and
has been moved to the `BaseModel` object accordingly.

Additionally, minor changes to how the `total_loss` (and its respective
`TensorFlow` operation) is built / calculated. The checks for the
various different 'types' of loss functions has been delegated to each
of the models' `calc_loss` methods.


Former-commit-id: 6eb76852dceb61574c7486e7c8af983ce8993386