Mod initial (MD) momenta by 2pi for `GaugeModel`
Explicitly, when building the `Dynamics` object to be used in a
`GaugeModel` instance, we want our randomly initialized momenta to be
contained in the interval [0, 2pi), so we take `tf.mod(v_init, 2 *
np.pi)` at the beginning of each MD trajectory (changes in
`_transition_forward(...)` and `_transition_backward(...)` methods of
`Dynamics` object).

The input values of the 'position' variables are taken modulo 2pi as
well at the beginning of each MD trajectory (changes in
`apply_transition(...)` method of the `Dynamics` object).


Former-commit-id: 675810218a44fd41a8b7e97e817f78180fa185d1