Includes remainder of changes for dEnergy plots
Specifically, includes new `EnergyPlotter` object in
`plotters/gauge_model_plotter.py` that plots time series and histograms
for each of the relevant energy differences.

Additionally, modified the `_create_energy_summaries` method in
`loggers/summary_utils.py` to be compatible the redefined `energies`
collection which contains all of the relevant operations for calculating
the energy differences.

Changes to `GaugeModelRunner` in `runners/gauge_model_runner.py`,
including the relevant method for running the energy ops when running
inference on a trained model.


Former-commit-id: 04a0873240b4ebdfa83869415a19c2065f6faac4