Creates `--zero_masks` flag for zeroed binary mask
Explicitly, instead of using a binary mask with half of its entries
equal to zero and half equal to one for each sub-update of `x`, this
uses a binary mask of all zeros (and complementary mask of all ones) for
each sub-update of `x`.

This may be useful for helping to further debug the source of the
plaquette bias.


Former-commit-id: 58dcae9048ed2da7ce21597bbccb646c110b53d6