Includes graph-independent inference code.
Specifically, creates `dynamics/dynamics_np.py`, containing the
`DynamicsRunner` object to perform inference without the need for the
tensorflow graph.

This is done by loading in the weights from the trained model and
calculating each of the S, Q, T functions imperatively in numpy.

NOTE: New requirement `jax` for performing automatic differentiation
(needed for calculating the gradient of the potential).