Fixes critical bug related to numpy inference.
For some reason, calling the `get_weights()` method on a given layer in
the model returned the weights matrix correctly but ALWAYS returned a
bias vector that was identically equal to zero.

This can be fixed by instead running:

```
weights, bias = sess.run(layer.weights)
```

which will then return the correct values.

I believe this may be an internal bug in tensorflow/keras, but further
testing on newer versions of tensorflow is needed prior to submitting an
issue on the tensorflow github repo.


Former-commit-id: cc6787995e3c672962bd0a5688b6619b1eafed69