Changes to how training is initialized
Specifically, introduces `--restore_from=/path/to/logdir` as an
additional flag to `train.py`.

When `restore_from` is specified, we load the saved networks from the
specified directory and restore the `GaugeDynamics` object from the most
recently saved checkpoint.

Additional checks are implemented to ensure that the networks are
compatible and a new directory is created if the annealing schedules
differ.


Former-commit-id: 75155ed727ca3ed8ac1b34ceedcf44a2e0206fcd